# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/carros": {
            "get": {
                "tags": [
                    "Carro"
                ],
                "summary": "Retornar a listagem de todos os carros",
                "description": "Get all Carros",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Carro"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/carros/pesquisar": {
            "post": {
                "tags": [
                    "Carro"
                ],
                "summary": "Pesquisa pelos Modelos de Carros",
                "description": "Store Carro",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Carro that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Carro"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Carro"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Carro": {
            "required": [
                ""
            ],
            "properties": {
                "modelo": {
                    "description": "modelo",
                    "type": "string"
                },
                "marca": {
                    "description": "marca",
                    "type": "string"
                },
                "tipo_de_veiculo": {
                    "description": "tipo_de_veiculo",
                    "type": "string"
                },
                "ano": {
                    "description": "ano",
                    "type": "string"
                },
                "potencia_do_motor": {
                    "description": "potencia_do_motor",
                    "type": "string"
                },
                "quilometragem": {
                    "description": "quilometragem",
                    "type": "string"
                },
                "combustivel": {
                    "description": "combustivel",
                    "type": "string"
                },
                "cambio": {
                    "description": "cambio",
                    "type": "string"
                },
                "direcao": {
                    "description": "direcao",
                    "type": "string"
                },
                "cor": {
                    "description": "cor",
                    "type": "string"
                },
                "portas": {
                    "description": "portas",
                    "type": "string"
                },
                "preco": {
                    "description": "preco",
                    "type": "string"
                },
                "final_de_placa": {
                    "description": "final_de_placa",
                    "type": "string"
                },
                "cep": {
                    "description": "cep",
                    "type": "string"
                },
                "municipio": {
                    "description": "municipio",
                    "type": "string"
                },
                "bairro": {
                    "description": "bairro",
                    "type": "string"
                }
            }
        }
    }
}